move test location
